Remove locales from Microsoft Docs URLs in Readme
Went to read this readme today and ended up on the Spanish version of the docs page for the Storage Account tutorial. Taking out the locale from the URL will make docs.microsoft automatically pick the most appropriate locale for your web request, so I propose that as the change here.